requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
celery
redis==4.3.1
djangorestframework
django-redis
django-celery-beat
pyyaml
uritemplate
django-rest-swagger